Split mobile session terminal pane
