+++ mobile/src/session/mobile-session-route-helpers.ts
@@ -1,0 +1,35 @@
+import type { TerminalModes } from '../terminal/TerminalWebView'
+import type { ConnectionState } from '../transport/types'
+
+export const MOBILE_SESSION_STATUS_LABELS: Record<ConnectionState, string> = {
+  connecting: 'Connecting',
+  handshaking: 'Securing',
+  connected: 'Connected',
+  disconnected: 'Disconnected',
+  reconnecting: 'Reconnecting',
+  'auth-failed': 'Auth failed'
+}
+
+export const TERMINAL_GESTURE_INPUT_BUCKET_CAPACITY = 64
+export const TERMINAL_GESTURE_INPUT_REFILL_PER_SECOND = 120
+export const TERMINAL_GESTURE_INPUT_FLUSH_DELAY_MS = 16
+export const TERMINAL_GESTURE_INPUT_MAX_PENDING_SEQUENCES = 32
+export const TERMINAL_GESTURE_INPUT_MAX_QUEUE_AGE_MS = 250
+
+export function isFileExistsErrorMessage(message: string): boolean {
+  const normalized = message.toLowerCase()
+  return normalized.includes('eexist') || normalized.includes('already exists')
+}
+
+export function getRepoIdFromMobileWorktreeId(id: string): string {
+  // Why: mobile cannot import desktop shared modules in its standalone tsc run,
+  // but the runtime worktree id wire format is still `${repoId}::${path}`.
+  const separatorIdx = id.indexOf('::')
+  return separatorIdx === -1 ? id : id.slice(0, separatorIdx)
+}
+
+export function isGestureMouseTrackingMode(
+  mode: TerminalModes['mouseTrackingMode'] | undefined
+): boolean {
+  return mode === 'x10' || mode === 'vt200' || mode === 'drag' || mode === 'any'
+}
